+++ fizzbuzz/src/main/java/edu/uagrm/core/FizzBuzz.java
@@ -1,4 +1,4 @@
-package edu.uagrm.main;
+package edu.uagrm.core;
 
 public interface FizzBuzz {
