+++ fizzbuzz/src/main/java/edu/uagrm/core/FizzBuzzType.java
@@ -1,4 +1,4 @@
-package edu.uagrm.main;
+package edu.uagrm.core;
 
 public enum FizzBuzzType {
   FIZZ, BUZZ, FIZZBUZZ, NUMBER
