+++ fizzbuzz/src/test/java/edu/uagrm/main/FizzBuzzTest.java
@@ -2,6 +2,7 @@ package edu.uagrm.main;
 
 import static org.junit.Assert.assertEquals;
 
+import edu.uagrm.core.FizzBuzz;
 import org.junit.Test;
 
 public class FizzBuzzTest {
